+++ pkg/client-sdk/store/inmemory/config.go
@@ -1,0 +1,1 @@
+package inmemory
