+++ pkg/client-sdk/store/inmemory/config_wasm.go
@@ -1,0 +1,1 @@
+package inmemory
