+++ pkg/client-sdk/store/inmemory/wallet_wasm.go
@@ -1,0 +1,1 @@
+package inmemory
